Timeouts are ints
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1158335 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/http11/Http11AprProcessor.java
@@ -182,7 +182,7 @@ public class Http11AprProcessor extends AbstractHttp11Processor<Long> {
         keepAlive = true;
         comet = false;
 
-        long soTimeout = endpoint.getSoTimeout();
+        int soTimeout = endpoint.getSoTimeout();
 
         if (disableKeepAlive()) {
             socketWrapper.setKeepAliveLeft(0);
